coap: split transport modules
